273 verified
git-svn-id: http://tradelink.googlecode.com/svn/trunk@2010 ccc6a032-7b47-0410-9fd8-1f29c74ce1d2

+++ BrokerServers/TradeLibFast/TLOrder.h
@@ -7,7 +7,7 @@ namespace TradeLibFast
 	public:
 		TLOrder(void);
 		~TLOrder(void);
-		unsigned int id;
+		long long id;
 		int size;
 		double price;
 		double stop;
